+++ src/primaite/__init__.py
@@ -16,6 +16,9 @@ from platformdirs import PlatformDirs
 with open(Path(__file__).parent.resolve() / "VERSION", "r") as file:
     __version__ = file.readline().strip()
 
+_PRIMAITE_ROOT: Path = Path(__file__).parent
+# TODO: Remove once we integrate the simulation into PrimAITE and it uses the primaite session path
+
 
 class _PrimaitePaths:
     """