+++ tests/unit_tests/_primaite/_simulator/_network/_transmission/test_network_layer.py
@@ -1,24 +1,24 @@
 import pytest
 
-from primaite.simulator.network.transmission.network_layer import ICMPHeader, ICMPType
+from primaite.simulator.network.transmission.network_layer import ICMPPacket, ICMPType
 
 
 def test_icmp_minimal_header_creation():
-    """Checks the minimal ICMPHeader (ping 1 request) creation using default values."""
-    ping = ICMPHeader()
+    """Checks the minimal ICMPPacket (ping 1 request) creation using default values."""
+    ping = ICMPPacket()
 
     assert ping.icmp_type == ICMPType.ECHO_REQUEST
     assert ping.icmp_code == 0
     assert ping.identifier
-    assert ping.sequence == 1
+    assert ping.sequence == 0
 
 
 def test_valid_icmp_type_code_pairing():
-    """Tests ICMPHeader creation with valid type and code pairing."""
-    assert ICMPHeader(icmp_type=ICMPType.DESTINATION_UNREACHABLE, icmp_code=6)
+    """Tests ICMPPacket creation with valid type and code pairing."""
+    assert ICMPPacket(icmp_type=ICMPType.DESTINATION_UNREACHABLE, icmp_code=6)
 
 
 def test_invalid_icmp_type_code_pairing():
-    """Tests ICMPHeader creation fails with invalid type and code pairing."""
+    """Tests ICMPPacket creation fails with invalid type and code pairing."""
     with pytest.raises(ValueError):
-        assert ICMPHeader(icmp_type=ICMPType.DESTINATION_UNREACHABLE, icmp_code=16)
+        assert ICMPPacket(icmp_type=ICMPType.DESTINATION_UNREACHABLE, icmp_code=16)